# .docs/swagger.yaml
basePath: /
definitions:
...
  http.AgentChatRequest:
    properties:
      friendID:
        example: 9b02ce54-4f42-4a8b-a539-5b53a6e37e63
        type: string
      message:
        example: Ela prefere experiencias ou objetos?
        type: string
    type: object
  http.AgentFinalizeRequest:
    properties:
      friendID:
        example: 9b02ce54-4f42-4a8b-a539-5b53a6e37e63
        type: string
    type: object
  http.ErrorResponse:
    properties:
      error:
        example: birthDate must be in format YYYY-MM-DD
        type: string
    type: object
...
  http.SuggestionChatRequest:
    properties:
      giftID:
        example: 77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0
        type: string
      message:
        example: Pode sugerir uma opcao mais economica?
        type: string
    type: object
  http.SuggestionCreateRequest:
    properties:
      occasionDetails:
        example: Aniversario em 2026-08-15, prefere experiencias ao ar livre
        type: string
      reminderID:
        example: d8c8efdf-c52f-4d6b-8e2e-b83f78de4f77
        type: string
    type: object
  http.SuggestionFinalizeRequest:
    properties:
      giftID:
        example: 77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para sugestões de presentes personalizadas.
  title: Gift App API
  version: "1.0"
paths:
...
  /profiles/{friendId}/suggestions:
    post:
      consumes:
      - application/json
      description: 'Exemplo de payload: {"occasionDetails":"Aniversario em 2026-08-15,
        prefere experiencias ao ar livre","reminderID":"d8c8efdf-c52f-4d6b-8e2e-b83f78de4f77"}.'
      parameters:
      - description: ID do friend
        in: path
        name: friendId
        required: true
        type: string
      - description: Dados da ocasião
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.SuggestionCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Criar sugestão inicial para friend
      tags:
      - suggestion-agent
  /profiles/agent/chat:
    post:
      consumes:
      - application/json
      description: 'Exemplo de payload: {"friendID":"9b02ce54-4f42-4a8b-a539-5b53a6e37e63","message":"Ela
        prefere experiencias ou objetos?"}.'
      parameters:
      - description: Payload do chat
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.AgentChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Conversar com profile agent
      tags:
      - profile-agent
  /profiles/agent/finalize:
    post:
      consumes:
      - application/json
      description: 'Exemplo de payload: {"friendID":"9b02ce54-4f42-4a8b-a539-5b53a6e37e63"}.'
      parameters:
      - description: Payload de finalização
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.AgentFinalizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Finalizar sessão de profile agent
      tags:
      - profile-agent
  /profiles/agent/session/{friendId}:
    delete:
      parameters:
      - description: ID do friend
        in: path
        name: friendId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Remover sessão do profile agent
      tags:
      - profile-agent
...
  /suggestions/agent/chat:
    post:
      consumes:
      - application/json
      description: 'Exemplo de payload: {"giftID":"77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0","message":"Pode
        sugerir uma opcao mais economica?"}.'
      parameters:
      - description: Payload do chat
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.SuggestionChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Conversar para refinar sugestão por gift
      tags:
      - suggestion-agent
  /suggestions/agent/finalize:
    post:
      consumes:
      - application/json
      description: 'Exemplo de payload: {"giftID":"77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0"}.'
      parameters:
      - description: Payload de finalização
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.SuggestionFinalizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Finalizar refinamento de sugestão por gift
      tags:
      - suggestion-agent
...
swagger: "2.0"